Refactor metrix source layout
